Watch Jasmine specs and source files
Use grunt-contrib-watch to watch specs and source files. Automatically
triggers a browersify build.
