+++ package.json
@@ -4,7 +4,7 @@
   "description": "A component LWleiterbox to be used in a vocabulary learning program",
   "main": "index.js",
   "scripts": {
-    "test": "node scripts/test.js"
+    "test": "grunt watch"
   },
   "repository": {
     "type": "git",
@@ -30,6 +30,8 @@
   "dependencies": {},
   "devDependencies": {
     "browserify": "^13.1.1",
+    "grunt": "^1.0.1",
+    "grunt-contrib-watch": "^1.0.0",
     "jasmine": "^2.5.2",
     "jasmine-expect": "^3.0.1"
   }
